% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC19043.R
\name{akde}
\alias{akde}
\title{An adaptive KDE}
\usage{
akde(x, h, kernel = "gaussian")
}
\arguments{
\item{x}{the data from which the estimate is to be computed.}

\item{h}{the bandwidth function to be used.}

\item{kernel}{a character string giving the smoothing kernel to be used, with default "gaussian".}
}
\value{
x  data points - same as input.

kernel name of kernel to use.

hx the bandwidth function - same as input.

h the bandwidth value at x to use.

fhat the estimated density function.
}
\description{
An adaptive kernel density estimator using R.
}
\examples{
\dontrun{
x<-rnorm(2000)
h<-function(u){(10+u^4)^(-0.6)}
fhat<-akde(x,h,kernel="gaussian")
plot(x,fhat$fhat(x),cex=0.5)
}
}
