\name{SC19043-package}
\alias{SC19043-package}
\alias{SC19043}
\docType{package}
\title{
  A R package made by SC19043
}
\description{
  The function _akde_ is used to solve the adaptive kernel density estimator problem.
  The function _est.knn_ is  to estimator density with k-Nearest Neighbor method.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Renhao Zou

Maintainer: SC19043 <1144573191@qq.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
